Extract class into file
