+++ toolkit/sharedlib/src/main/java/com/arcgismaps/toolkit/ui/expandablecard/ExpandableCarsState.kt
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2024 Esri
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.arcgismaps.toolkit.ui.expandablecard
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.Saver
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+
+/**
+ * State object that can be hoisted to control the [ExpandableCard].
+ *
+ * In most cases, this will be created using [rememberExpandableCardState].
+ * @since 200.6.0
+ */
+class ExpandableCardState constructor(initialExpandedSttate: Boolean) {
+    /**
+     * The expanded state of the [ExpandableCard].
+     * @since 200.6.0
+     */
+    var isExpanded by mutableStateOf(initialExpandedSttate)
+}
+
+/**
+ * Remember the state of [ExpandableCard].
+ *
+ * @param isExpanded the initial expanded state
+ * @since 200.6.0
+ */
+@Composable
+fun rememberExpandableCardState(
+    isExpanded: Boolean = true
+): ExpandableCardState {
+    return rememberSaveable(
+        saver = Saver(
+            save = { it.isExpanded},
+            restore = { ExpandableCardState(it) }
+        )
+    ) {
+        ExpandableCardState(isExpanded)
+    }
+}
